Move SVG image types to node::svg; parse XML once for tree and dimensions
