+++ packages/blitz-dom/src/node/mod.rs
@@ -7,6 +7,8 @@ mod element;
 mod node;
 pub(crate) mod scrollbar;
 mod stylo_data;
+#[cfg(feature = "svg")]
+mod svg;
 mod text;
 
 pub use attributes::{Attribute, Attributes};
@@ -19,8 +21,8 @@ pub use element::{
     ListItemLayoutPosition, Marker, RasterImageData, SpecialElementData, SpecialElementType,
     Status,
 };
-#[cfg(feature = "svg")]
-pub use element::{SvgImageData, SvgIntrinsicDimensions};
 pub use node::*;
 pub use scrollbar::{ScrollbarColor, ScrollbarRef, ScrollbarWidth};
+#[cfg(feature = "svg")]
+pub use svg::{SvgImageData, SvgIntrinsicDimensions};
 pub use text::{GeneratedTextInputEvent, TextBrush, TextInputData, TextLayout};
